move enum builder
