+++ libs/util/code-gen-utils/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./lib/util-code-gen-utils";
+export * from "./lib/enum-builder";
