+++ libs/util/code-gen-utils/src/lib/enum-builder.ts
