+++ package.json
@@ -240,6 +240,7 @@
     "react-is": "^18.2.0",
     "react-test-renderer": "^18.2.0",
     "react-transition-group": "^4.4.2",
+    "recast": "^0.20.5",
     "semver": "^7.3.7",
     "source-map-support": "^0.5.21",
     "supertest": "^4.0.2",
